Board #1

+++ mysite02/src/main/java/com/douzone/mysite/vo/BoardVo.java
@@ -4,7 +4,7 @@ public class BoardVo {
 	private Long no;
 	private String title;
 	private String contents;
-	private Long hit;
+	private int hit;
 	private String reg_date;
 	private Long group_no;
 	private Long order_no;
@@ -37,10 +37,11 @@ public class BoardVo {
 	public void setContents(String contents) {
 		this.contents = contents;
 	}
-	public Long getHit() {
+	
+	public int getHit() {
 		return hit;
 	}
-	public void setHit(Long hit) {
+	public void setHit(int hit) {
 		this.hit = hit;
 	}
 	public String getReg_date() {
@@ -79,6 +80,5 @@ public class BoardVo {
 				+ reg_date + ", group_no=" + group_no + ", order_no=" + order_no + ", depth=" + depth + ", user_no="
 				+ user_no + ", name=" + name + "]";
 	}
-
 
 }
